Another Rebuild. Ready for Cast
I have removed all the use of GMNotes and Names and Descriptions and switched entirely to TAGS. I feel like I have been futsing about and have NO plans to improve this... I'm just going to run with it. I'm ready to do the cast code and finsh this modual of the mod.